proyecto/proyecto/requirements.txt
Django==2.0
django-cors-headers==3.4.0
django-simple-history==2.11.0
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
psycopg2-binary==2.8.5
PyJWT==1.7.1
python-decouple==3.3
pytz==2020.1
six==1.15.0
